move Printer into its own file
